create event : code clean up
