+++ app/src/main/java/com/example/c4q/capstone/userinterface/events/createevent/CreateEventAddVenueFragment.java
@@ -38,7 +38,6 @@ public class CreateEventAddVenueFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        //CurrentUser.getInstance().
     }
