+++ app/src/main/java/com/example/c4q/capstone/userinterface/events/createevent/CreateEventInviteFragment.java
@@ -36,9 +36,6 @@ public class CreateEventInviteFragment extends Fragment {
         return fragment;
     }
 
-
-
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
