+++ app/src/main/java/com/example/c4q/capstone/userinterface/events/eventfragments/SingleEventFragment.java
@@ -65,8 +65,6 @@ public class SingleEventFragment extends Fragment {
         setVoteClick();
         showHideMore();
 
-
-
         return rootView;
     }
